Add a Daemon for the Webui
Added a Daemon class and a lightshowpi_daemon (sub-class of Daemon).
To see the usage of the daemon:lightshowpi_daemon.py start|stop|restart [Optional] port
This allows the webui to be started from a terminal, then the terminal
closed. Also corrected an indention error in weapp.py. and synchonized_lights.py
Within webapp.py move the variable declaration of the "app" variable up to
just after the declaration of the urls. Do this to allow the import the
variable into the Daemon file and run it from there. The pid file of the
Daemon is created in the "env" directory.